feat(waiting): WaitingService를 인터페이스와 구현체로 분리 (#66)
